First Release
This release contains the first page of the tome. It has neat colored text, and a button at the bottom of the page that allows you to teleport back to your main home.

Command used for teleporting home is /trigger home

The next update will include pages to set homes up to your allowed limit, as well as pages to recall to a specific location that isn't your main home.

Suggestions are always appreciated!